requirements.txt
mediapipe==0.10.9
opencv-python==4.9.0.80
streamlit==1.30.0
numpy==1.26.3
ffmpeg-python==0.2.0
uuid==1.30
pillow==10.2.0